seperate coin card widget
