+++ frontend/lib/views/coins_view.dart
@@ -2,6 +2,7 @@ import 'dart:convert';
 
 import 'package:flutter/material.dart';
 import 'package:flutter/services.dart';
+import 'package:frontend/widgets/coin_card.dart';
 
 /// This is CoinsView, for buying coins from store
 class CoinsView extends StatefulWidget {
@@ -66,84 +67,6 @@ class _CoinsViewState extends State<CoinsView> {
             ],
           ),
         ),
-      ),
-    );
-  }
-}
-
-class CoinCard extends StatelessWidget {
-  final String label;
-  final int coinCount;
-  final double price;
-  final String imagePath = 'lib/assets/images/coin-zuggled.png';
-
-  const CoinCard({
-    super.key,
-    required this.label,
-    required this.coinCount,
-    required this.price,
-  });
-
-  @override
-  Widget build(BuildContext context) {
-    return Card(
-      color: Theme.of(context).colorScheme.surface,
-      shape: RoundedRectangleBorder(
-        borderRadius: BorderRadius.circular(16),
-        side: BorderSide(color: Theme.of(context).colorScheme.outlineVariant),
-      ),
-      elevation: 0,
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: [
-          Container(
-            height: 110,
-            width: double.infinity,
-            decoration: const BoxDecoration(
-              color: Color(0xFFF0ECF9),
-              borderRadius: BorderRadius.vertical(top: Radius.circular(16)),
-            ),
-            child: ClipRRect(
-              borderRadius: const BorderRadius.vertical(top: Radius.circular(16)),
-              child: Image.asset(
-                imagePath,
-                fit: BoxFit.contain,
-              ),
-            ),
-          ),
-          const SizedBox(height: 10),
-          Text(
-            label,
-            style: const TextStyle(fontSize: 15, fontWeight: FontWeight.bold),
-            textAlign: TextAlign.center,
-          ),
-          const SizedBox(height: 2),
-          Text(
-            '$coinCount Coins',
-            style: const TextStyle(fontSize: 14),
-          ),
-          const SizedBox(height: 8),
-          TextButton(
-            onPressed: () {
-              // Handle top-up logic here (e.g., open payment gateway)
-            },
-            style: TextButton.styleFrom(
-              backgroundColor: Theme.of(context).colorScheme.primary,
-              padding: const EdgeInsets.symmetric(horizontal: 16, vertical: 8),
-              shape: RoundedRectangleBorder(
-                borderRadius: BorderRadius.circular(24),
-              ),
-            ),
-            child: Text(
-              '₹${price.toStringAsFixed(0)}',
-              style: const TextStyle(
-                color: Colors.white,
-                fontWeight: FontWeight.bold,
-              ),
-            ),
-          ),
-          const SizedBox(height: 10),
-        ],
       ),
     );
   }
